+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v1.3
 * BUGFIX - Fixed bug with --max\_fragment\_size where reverse-stranded properly paired reads where still being used as a fragment since the template length reported is negative
 * FEATURE - Adding --remove\_multimapped as an option.  Enabling it removes any reads mapping to multiple regions in the genome.
+* FEATURE - Adding --chunk_size as an option.  Sets the number of validated reads/fragments that are read into memory before processing overlaps with features.
+* Changed the custom BAM record IntervalCollection to simply read 'R' or 'F' as the metadata instead of "read" or "fragment".  This should hopefully be more efficient as Char types are used instead of String types.
 
 ## v1.2
 * FEATURE - Added two new output columns.  1) Nonoverlapping feature length. 2) TPM (in scientific notation)